% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSummary.R
\name{dataSummary}
\alias{dataSummary}
\title{Statistical Summaries of a Numeric Data Set}
\usage{
dataSummary(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{A numeric vector containing the data}

\item{na.rm}{Should \code{NA} be removed?  Default value \code{TRUE}.}
}
\value{
A named vector of numerical summaries and the number of \code{NA}:
\describe{
\item{Mean}{The mean of the data.}
\item{Standard Deviation}{The standard deviation of the data.}
\item{NAs}{The number of missing data}
}
}
\description{
This function provides statistical summaries of
a numeric data set.
It calls the \code{statsCalculate} function.
It also plots a histogram using the \code{ggplot2} package.
}
\examples{
x <- rnorm(100)
dataSummary(x)

}
\author{
Luciana Dalla Valle \email{luciana.dallavalle@plymouth.ac.uk}, Julian Stander \email{J.Stander@plymouth.ac.uk}
}
