% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstRpack-package.R
\docType{package}
\name{myfirstRpack-package}
\alias{myfirstRpack}
\alias{myfirstRpack-package}
\title{myfirstRpack: A Package to Provide Numerical and Graphical Summaries of a Data Set}
\description{
The package provides numerical and graphical summaries of a data set.
}
\author{
\strong{Maintainer}: Luciana Dalla Valle \email{luciana.dallavalle@plymouth.ac.uk}

Authors:
\itemize{
  \item Julian Stander \email{J.Stander@plymouth.ac.uk}
}

}
\keyword{internal}
