% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsCalculate.R
\name{statsCalculate}
\alias{statsCalculate}
\title{Mean and Standard Deviation of a Numeric Data Set}
\usage{
statsCalculate(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector containing the data.}

\item{na.rm}{Should \code{NA} be removed?  Default value \code{TRUE}.}
}
\value{
A named vector of numerical summaries and the number of \code{NA}:
\describe{
\item{Mean}{The mean of the data.}
\item{Standard Deviation}{The standard deviation of the data.}
}
}
\description{
This function provides two statistical summaries of
a numeric data set.
These comprise a measure of location in the form of a mean
and a measure of spread in the form of a standard deviation.
}
\examples{
x <- rnorm(100)
statsCalculate(x)

}
\author{
Luciana Dalla Valle \email{luciana.dallavalle@plymouth.ac.uk}, Julian Stander \email{J.Stander@plymouth.ac.uk}
}
